Adding test and documentation Also deleting teacher's project template
Done by Timothée PAVAGEAU
But unable to push

+++ src/main/scala/Event.scala
@@ -22,8 +22,8 @@ case class Event(
                   lastname: String,
                   firstname: String,
                   age: Int,
-                  latitude: Int,
-                  longitude: Int,
+                  latitude: Double,
+                  longitude: Double,
                   pain: Int,
                   need: String,
                )
